chore: split TemporaryBucket into its own class
